state has to be set

+++ webapp/src/main/java/rocks/metaldetector/security/filter/OAuth2AuthorizationCodeLoginFilter.java
@@ -34,16 +34,19 @@ public class OAuth2AuthorizationCodeLoginFilter extends OncePerRequestFilter {
   @Override
   protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
     String state = request.getParameter("state");
-    try {
-      AbstractUserEntity user = authorizationStateService.findUserByState(state);
-      UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities());
-      authentication.setDetails(authenticationDetailsSource.buildDetails(request));
-      SecurityContextHolder.getContext().setAuthentication(authentication);
-      authorizationStateService.deleteByState(state);
-    }
-    catch (Exception e) {
-      log.error("Cannot authenticate user returning from authorization code flow", e);
-      response.sendError(SC_FORBIDDEN);
+    log.info(state == null ? "state is null" : "state: " + state);
+    if (state != null && !state.isBlank()) {
+      try {
+        AbstractUserEntity user = authorizationStateService.findUserByState(state);
+        UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities());
+        authentication.setDetails(authenticationDetailsSource.buildDetails(request));
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+        authorizationStateService.deleteByState(state);
+      }
+      catch (Exception e) {
+        log.error("Cannot authenticate user returning from authorization code flow", e);
+        response.sendError(SC_FORBIDDEN);
+      }
     }
 
     filterChain.doFilter(request, response);
